backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
psycopg[binary]==3.2.4
alembic==1.13.0
pydantic[email]==2.9.0
pydantic-settings==2.5.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
bcrypt==4.0.1
easyocr==1.7.2
Pillow==11.1.0
httpx==0.27.0
python-multipart==0.0.9
